Renamed file from segmentes to segmented, Fixed Unit test
