+++ src/kirby/components/chip/chip.component.ts
@@ -7,7 +7,7 @@ import { Component, ViewEncapsulation, Input } from '@angular/core';
   encapsulation: ViewEncapsulation.None,
 })
 export class ChipComponent {
-  @Input() label: String;
+  @Input() label: string;
   @Input() isSelected: boolean;
   constructor() {}
 }
