+++ src/kirby/components/segment/segmented-chip-control/segmented-chip-control.component.html
@@ -1,0 +1,5 @@
+<div class="scrollable">
+    <div *ngFor="let item of items" (click)="segmentedChipSelect(item)">
+        <kirby-chip [label]="item" [isSelected]="item.toLowerCase() === selectedItem.toLowerCase()"></kirby-chip>
+    </div>
+</div>
