+++ src/kirby/components/segment/segmented-chip-control/segmented-chip-control.component.tns.scss
