+++ src/kirby/components/segment/segmented-chip-control/segmented-chip-control.component.ts
@@ -2,8 +2,8 @@ import { Component, Input, OnInit, Output, EventEmitter } from '@angular/core';
 
 @Component({
   selector: 'kirby-segmented-chip-control',
-  templateUrl: './segmentes-chip-control.component.html',
-  styleUrls: ['./segmentes-chip-control.component.scss'],
+  templateUrl: './segmented-chip-control.component.html',
+  styleUrls: ['./segmented-chip-control.component.scss'],
 })
 class SegmentedChipControlComponent implements OnInit {
   @Input() items: string[];
