--- src/kirby/components/segment/segmentes-chip-control/segmentes-chip-control.component.html
@@ -1,5 +1,0 @@
-   <div class="scrollable">
-            <div *ngFor="let item of items" (click)="segmentedChipSelect(item)">
-                <kirby-chip [label]="item" [isSelected]="item.toLowerCase() === selectedItem.toLowerCase()"></kirby-chip>
-            </div>
-    </div>
